moved all triton tests into runtime triton folder
