+++ tests/unit/systems/dag/runtimes/triton/ops/implicit/__init__.py
@@ -13,3 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+
+import pytest
+
+pytest.importorskip("implicit")
